fix: review entity aligned with db model

+++ src/main/java/br/com/ragro/domain/Review.java
@@ -36,7 +36,7 @@ public class Review {
     private Producer farmer;
 
     @Column(name = "rating", nullable = false)
-    private Integer rating;
+    private Short rating;
 
     @Column(name = "comment", columnDefinition = "text")
     private String comment;
